test(lsm): consolidate negative-cache test/bench files into lsm_test/lsm_bench_test
The engine/lsm/negative_cache.go source was deleted in 56b72a2e (the
cache itself moved to engine/slab/negativecache/), leaving the
companion negative_cache_test.go and negative_cache_bench_test.go as
orphan files that no longer mapped to any source file.

The pure cache semantics tests already live with the cache in
engine/slab/negativecache/cache_test.go (TestCacheRememberHasInvalidate,
TestCacheClearWipesEverything, TestCacheSnapshotKeysDropsStaleEntries,
TestCacheConcurrentRememberHas, TestPersistenceRoundTrip,
TestPersistenceMissingFileIsClean). The 5 lsm-side tests are
genuinely LSM-integration — they exercise lsm.Get's miss-path Remember
hook, lsm.SetBatchGroup's Invalidate, the range-tombstone
short-circuit in lsm.Get, lsm.clearNegativeCache, and the
NegativeCachePersistent option flow through Open/Close — so they
cannot move to slab/negativecache (which has no LSM in scope).

Fold them into lsm_test.go / lsm_bench_test.go so we stop having
orphan files. lsm_test.go gains a labeled section break that points
new readers to slab/negativecache for the underlying cache contract.

Files
─────
  engine/lsm/lsm_test.go
    + 5 negative-cache integration tests + walManagerHandle helper
  engine/lsm/lsm_bench_test.go
    + BenchmarkShardedNegativeCache + math/rand import
  engine/lsm/negative_cache_test.go        DELETED
  engine/lsm/negative_cache_bench_test.go  DELETED

Verification
  go test -race ./engine/... .             green (negative tests run)
  go test -bench BenchmarkShardedNegativeCache ./engine/lsm/  passes
  make lint (golangci-lint v2.9.0)         0 issues
